+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Delta-Coverage Gradle plugin Changelog
 
+## 2.2.0
+
+### Added
+- Added native Git support #53.
+
+### Fixed
+- Fixed diff generation by JGit client. Now it ignores whitespaces.
+- Fixed broken console report layout when class name is too long #92.
+
+### Changed
+- HTML report link is printed after console report.
+
+
 ## 2.1.0
 
 ### Added